Added auto clicker and Directx workaround
pynput uses a Win32API function called SendInput. The SendInput function will insert input events into the same queue as a hardware device but the events are marked with a LLMHF_INJECTED flag that can be detected by hooks and then filtered. So the DirectX workaround and key codes from Gautam-J